+++ src/components/primitives/index.ts
@@ -1,6 +1,7 @@
 export * from "./checkbox";
 export * from "./color";
 export * from "./date";
+export * from "./dateInput";
 export * from "./multiline";
 export * from "./number";
 export * from "./password";